Update gFEX input ByteStream decoder
Update gFEX input ByteStream decoder

+++ Event/xAOD/xAODTrigL1Calo/xAODTrigL1Calo/versions/gFexTower_v1.h
@@ -39,11 +39,11 @@ namespace xAOD {
       void setFpga(uint8_t); /// setter for the above
       
       /// get Energy Value
-      int16_t towerEt() const; /// getter for the energy value
+      int towerEt() const; /// getter for the energy value
       /// set Energy Value
-      void setTowerEt(int16_t); /// setter for the above
+      void setTowerEt(int); /// setter for the above
       /// set signed value of energy
-      int16_t signedEt(int16_t TowerEt) const; /// extract the signed value of the tower
+      int16_t signedEt(int TowerEt) const; /// extract the signed value of the tower
 
       /// Is gTower saturated?
       char isSaturated() const; /// getter for the saturation flag of gTower
@@ -52,8 +52,8 @@ namespace xAOD {
 
       void initialize(const uint8_t Eta,const uint8_t Phi);
       void initialize(const uint8_t Eta,const uint8_t Phi,
-                      const int16_t TowerEt,const uint8_t Fpga,
-				          const char IsSaturated);
+                      const int TowerEt,const uint8_t Fpga,
+				              const char IsSaturated);
 
     
     private:
